Update sane to the latest version and refresh lockfile
The version of sane currently being used pulls in other dependencies
that have known security issues. Update to the latest version of
sane to fix this. More info here:

https://github.com/pinterest/esprint/issues/92

Also refresh the lockfile (delete and regenerate) to grab the latest
versions of all dependencies.

(An apparent side-effect of this is an increase in the pass rate of
the unit tests. A few are still failing though.)